充值列表value改为double

+++ app/src/main/java/com/xiaolian/amigo/data/network/model/wallet/RechargeDenominations.java
@@ -12,5 +12,5 @@ public class RechargeDenominations {
     private Integer activityType;
     private Double amount;
     private Long id;
-    private Long value;
+    private Double value;
 }
